exclusive file to read json map file
