+++ Assets/Scenes/Board1.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 68fb191aec2d31b4fa419ec809d5d2f6
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
