+++ Assets/Scripts/JSONMapReader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de99f1f0a32831f47922f6d592ef0ac5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
